refactor: Split meta/plain/common tests
