+++ tests/common/test_ramp_A_precise.py
