+++ tests/common/test_receiver.py
@@ -18,14 +18,6 @@ def test_exchange(alice, charlie, swap, wrapped_coins):
     assert wrapped_coins[0].balanceOf(alice) == 0
 
 
-def test_exchange_underlying(alice, charlie, swap, underlying_coins):
-    underlying_coins[1]._mint_for_testing(alice, 10**18)
-
-    swap.exchange_underlying(1, 0, 10**18, 0, charlie, {'from': alice})
-    assert underlying_coins[0].balanceOf(charlie) > 0
-    assert underlying_coins[0].balanceOf(alice) == 0
-
-
 def test_remove_liquidity(alice, swap, charlie, wrapped_coins, initial_amounts):
     initial_amount = swap.balanceOf(alice)
     withdraw_amount = initial_amount // 4
