+++ tests/common/test_remove_liquidity_one_coin.py
@@ -5,20 +5,6 @@ pytestmark = pytest.mark.usefixtures("add_initial_liquidity")
 
 
 @pytest.mark.parametrize("idx", range(2))
-def test_amount_received(chain, alice, swap, wrapped_coins, wrapped_decimals, base_pool, idx):
-    decimals = wrapped_decimals[idx]
-    wrapped = wrapped_coins[idx]
-
-    swap.remove_liquidity_one_coin(10**18, idx, 0, {'from': alice})
-
-    ideal = 10**decimals
-    if idx == 1:
-        ideal = 10**36 // base_pool.get_virtual_price()
-
-    assert ideal*0.99 <= wrapped.balanceOf(alice) <= ideal
-
-
-@pytest.mark.parametrize("idx", range(2))
 @pytest.mark.parametrize("divisor", [1, 5, 42])
 def test_lp_token_balance(alice, swap, idx, divisor):
     initial_amount = swap.balanceOf(alice)
