+++ tests/meta/test_exchange.py
@@ -33,15 +33,3 @@ def test_exchange(
     admin_fee = swap.admin_balances(receiving)
 
     assert expected_admin_fee / admin_fee == approx(1, rel=max(1e-3, 1/(expected_admin_fee-1.1)))
-
-
-@pytest.mark.parametrize("sending,receiving", [(0, 1), (1, 0)])
-def test_min_dy(bob, swap, wrapped_coins, sending, receiving, wrapped_decimals):
-    amount = 10**wrapped_decimals[sending]
-    wrapped_coins[sending]._mint_for_testing(bob, amount, {'from': bob})
-
-    min_dy = swap.get_dy(sending, receiving, amount)
-    swap.exchange(sending, receiving, amount, min_dy-1, {'from': bob})
-
-    received = wrapped_coins[receiving].balanceOf(bob)
-    assert abs(received - min_dy) <= 1
